requirements.txt
MouseInfo==0.1.3
numpy==1.18.4
pandas==1.0.3
Pillow==9.0.1
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
six==1.14.0
easygui==0.98.1
